requirements.txt
pyxdg
python-dateutil
numpy
scipy
matplotlib
nose
toml
